still 01

+++ src/Emitter.sol
@@ -1,23 +1,18 @@
-// UNLICENSED
+//SPDX-License-Identifier: Unlicense
 pragma solidity ^0.8.13;
 
 contract Emitter {
-    
-    event Blip(uint indexed count, string message);
+    event Blip(uint256 indexed count, string message);
 
-    uint counter;
+    uint256 counter;
 
     constructor() {
-        
         counter = 0;
-
     }
 
-    function TocToc(string memory message) public { 
-        
+    function TocToc(string memory message) public {
         emit Blip(counter, message);
 
         counter++;
     }
-
 }
